Fix: Correct Order Details Progress and Data Types
This commit introduces several fixes and minor enhancements across the order details, product details, and home screens.

The primary change is the correction of the order status progress bar, which now accurately reflects the `orderStatusPercentage` received from the API. Data types for pricing have also been updated to handle decimal values correctly.

**Key Changes:**

*   **Order Details Screen (`OrderDetailsFragment.kt`, `fragment_order_details.xml`):**
    *   The `LinearProgressIndicator` (now `orderProgress`) is now correctly updated with the `orderStatusPercentage` value from the `FetchOrderDetailResponse`.
    *   The `pricing` field in `GetOrderDetailsResponse` has been changed from `Int?` to `Double?` to correctly handle currency values.
    *   Minor layout cleanup and padding adjustments were made.

*   **API Response Models:**
    *   **`FetchOrderDetailResponse.kt`**: Added the `orderStatusPercentage` field to the response model to drive the progress bar.
    *   **`GetOrderDetailsResponse.kt`**: Changed `pricing` from `Int?` to `Double?`.

*   **Minor Fixes & Cleanups:**
    *   **`ProductDetailsFragment.kt`**: Simplified string construction in the `shareSellerProfile` function.
    *   **`fragment_home.xml`**: Changed the keyboard input type for the search bar to `textAutoComplete` for a better user experience.
    *   Removed extra whitespace from `ic_sound_off.xml` and `fragment_product_details.xml`.

+++ app/src/main/java/io/bidswipe/app/network/response/GetOrderDetailsResponse.kt
@@ -98,7 +98,7 @@ data class GetOrderDetailsResponse(
 			@SerializedName("images")
 			val images : List<String?>?,
 			@SerializedName("pricing")
-			val pricing : Int?,
+			val pricing : Double?,
 			@SerializedName("product_show")
 			val productShow : String?,
 			@SerializedName("purchased_quantity")
